requirements.txt
langgraph>=0.2.0
openai>=1.40.0
tavily-python>=0.5.0
requests>=2.31.0
python-dotenv>=1.0.1
typing-extensions>=4.9.0
